+++ lib/strong_migrations.rb
@@ -166,17 +166,13 @@ end",
     change_column_null_postgresql:
 "Setting NOT NULL on a column requires an AccessExclusiveLock,
 which is expensive on large tables. Instead, use a constraint and
-validate it in a separate migration with a more agreeable RowShareLock.
+validate it in a separate step with a more agreeable RowShareLock.
 
 class %{migration_name} < ActiveRecord::Migration%{migration_suffix}
-  def change
-    %{add_constraint_code}
-  end
-end
+  disable_ddl_transaction!
 
-class Validate%{migration_name} < ActiveRecord::Migration%{migration_suffix}
   def change
-    %{validate_constraint_code}
+    %{command}
   end
 end",