+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## 0.5.2 (unreleased)
 
 - Added `add_foreign_key_safely` method
-- Added `change_column_null_safely` method
+- Added `add_null_constraint_safely` and `remove_null_constraint_safely` methods
 - Added `statement_timeout` and `lock_timeout` functionality
 
 ## 0.5.1 (2019-12-17)