+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.5.2 (unreleased)
 
 - Adding a column with a non-null default value is safe in MariaDB 10.3.2+
+- Added MariaDB-specific check for `change_column_null`
 - Added `add_foreign_key_safely` method
 - Added `add_null_constraint_safely` and `remove_null_constraint_safely` methods
 - Added `statement_timeout` and `lock_timeout` functionality